+++ examples/gateway/dependency_injection.py
@@ -10,7 +10,7 @@ class MyDatabase:
         self.value = value
 
 bot = hikari.GatewayBot("...")
-client = arc.GatewayClient(bot)
+client = arc.GatewayClientBase(bot)
 
 # Create a new instance of 'MyDatabase'
 database = MyDatabase(value=0)