+++ examples/gateway/options.py
@@ -5,7 +5,7 @@ import arc
 # Related documentation guide: https://arc.hypergonial.com/guides/options
 
 bot = hikari.GatewayBot("...")
-client = arc.GatewayClient(bot)
+client = arc.GatewayClientBase(bot)
 
 
 @client.include
@@ -42,7 +42,7 @@ async def all_the_options(
 
 # Define an autocompletion callback
 # This can either return a list of the option type, or a list of hikari.CommandChoice
-async def provide_opts(data: arc.AutocompleteData[arc.GatewayClient, str]) -> list[str]:
+async def provide_opts(data: arc.AutocompleteData[arc.GatewayClientBase, str]) -> list[str]:
     if data.focused_value and len(data.focused_value) > 20:
         return ["That", "is", "so", "long!"]
     return ["Short", "is", "better!"]