+++ examples/rest/hooks.py
@@ -7,7 +7,7 @@ import arc
 # Related documentation guide: https://arc.hypergonial.com/guides/hooks
 
 bot = hikari.RESTBot("...")
-client = arc.RESTClient(bot)
+client = arc.RESTClientBase(bot)
 
 # Any function that takes a context as its first argument
 # and returns None or HookResult is a valid hook