+++ docs/changelog.md
@@ -9,8 +9,9 @@ hide:
 
 Here you can find all the changelogs for `hikari-arc`.
 
-## v0.4.1
+## Unreleased
 
+- Add `GatewayClientBase` and `RESTClientBase` to aid in creating custom client types. Examples on how to do this have also been added to the repository.
 - Fix `InteractionResponse.retrieve_message()` failing due to incorrect assertion.
 
 ## v0.4.0